fix(fala-4): BOM / routing / FactorySpec — 10 findings R06 (post-fix audit 16.07)
PF-R06-02  FactorySpec Review oznacza shelf life jako dziedziczony z itemu FG
           (bez duplikatu kolumny — wartość idzie joinem z items) + trasa do itemu
PF-R06-03  scrap % edytowalny (audytowany, V-TEC-11 ostrzega przy update,
           3. miejsce po przecinku jawnie odrzucane zamiast cicho zaokrąglane)
PF-R06-04  kolejność linii BOM — akcja move up/down z dwufazową renumeracją
PF-R06-05  kolejność operacji routingu — kontrolki ↑/↓, opNo renumerowane 1..N
PF-R06-06  kasowanie wersji routingu w statusie draft (guard + audyt + potwierdzenie)
PF-R06-07  selektor linii pokazuje site i zawęża po routings.site_id
PF-R06-08  edycja aktywnego BOM mówi prawdę: powstaje wersja in_review, nie draft
PF-R06-09  ułamkowy setup minutes działa całą ścieżką (migracja 523 numeric(18,6))
PF-R06-10  released/archived BOM nie pokazuje już wykonalnych kontrolek mutacji
PF-R06-11  snapshot BOM powstaje przy TWORZENIU WO, nie dopiero przy starcie

Cross-review (Codex) zgłosił 33 znaleziska; wszystkie zarbitrażowane i naprawione
w dwóch rundach. Najgroźniejsze:
- błąd snapshotu zostawiał OSIEROCONY nagłówek WO (withOrgContext commituje przy
  zwykłym return — raport toru twierdził, że cofa; nie cofał)
- post-check migracji 523 brał dowolny wiersz przez `limit 1` bez `order by`,
  a trigger z mig 496 blokuje UPDATE operacji zablokowanego routingu → deploy
  byłby rzutem monetą (potwierdzone dowodem na prodzie)
- site-RLS pozwalało guardowi kasowania routingu przeoczyć WO z innego site'u
- ścieżka ODCZYTU nadal gubiła precyzję setup_time_min (JSON.parse + Number)
- cały namespace i18n technical.bom.rowActions nie istniał w runtime bundlach

Defekty ZASTANE naprawione przy okazji:
- `Save version` gubił współprodukty i resetował yield do 100 %
- edycja linii bez operacji była zablokowana mimo że serwer dopuszcza null

Migracje: 523 (setup_time_min → numeric(18,6)), 524 (unique na bom_snapshots),
525 (SECURITY DEFINER na liczenie referencji routingu).
Każda PREPARE'owana na prodzie; 523 zweryfikowana 3× pod rząd (post-check tworzy
i wycofuje własny probe-routing, więc nie zależy od danych produkcyjnych).

Bramka: typecheck 0, build 0, core 0 regresji (+86 zielonych testów),
UI 0 regresji (suita UI nie była uruchamiana w poprzednich falach — `&&` w skrypcie).

+++ apps/web/app/[locale]/(app)/(modules)/technical/routings/_actions/create-routing.ts
@@ -10,8 +10,13 @@
  * are written in the SAME transaction (atomic), so a V-TEC failure mid-insert
  * leaves no partial routing.
  *
- * Supersede pattern (same as BOMs): a new routing starts at `status='draft'`. We
- * NEVER delete routing rows — versions are superseded, never hard-deleted.
+ * Supersede pattern (same as BOMs): a new routing starts at `status='draft'`.
+ * An approved, active or superseded routing is NEVER hard-deleted — those rows
+ * are the audit trail and are superseded instead. `draft` is the one exception
+ * (PF-R06-06, see delete-routing.ts): a draft was never approved, never produced
+ * against and carries no signature, so it is not part of that history and a
+ * mistakenly created one can be removed. Same line BOM draws with
+ * deleteBomVersion.
  */
 
 import { withOrgContext } from '../../../../../../../lib/auth/with-org-context';
@@ -137,7 +142,7 @@ export async function createRouting(rawInput: unknown): Promise<CreateRoutingRes
               setup_time_min, run_time_per_unit_sec, cost_per_hour, manufacturing_operation_name, crew, yield_pct, created_by)
            values
              (app.current_org_id(), $1::uuid, $2::integer, $3, $4, $5::uuid,
-              $6::integer, $7::numeric, $8::numeric, $9, $10::jsonb, $11::numeric, $12::uuid)`,
+              $6::numeric, $7::numeric, $8::numeric, $9, $10::jsonb, $11::numeric, $12::uuid)`,
           [
             routing.id,
             op.opNo,
